Apply agent changes
